# docs/api/openapi.public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpsNinja API",
    "version": "1.0.0",
    "description": "OpsNinja is a multi-tenant SaaS support and incident-management platform. This document describes the versioned, publicly supported tenant-facing REST surface. Internal endpoints (health, auth, admin, Jira integration) are excluded from this document.",
    "contact": {
      "name": "OpsNinja API Support",
      "url": "https://docs.opsninja.io",
      "email": "api-support@opsninja.io"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://opsninja.io/terms"
    }
  },
  "servers": [
    { "url": "https://api.opsninja.io", "description": "Production" },
    { "url": "https://api.staging.opsninja.io", "description": "Staging" },
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
...
  "security": [{ "StaffBearer": [] }],
  "paths": {
    "/api/v1/portal/tickets": {
      "get": {
        "operationId": "listPortalTickets",
        "summary": "List portal tickets with cursor pagination",
        "tags": ["portal-tickets"],
        "security": [{ "PortalBearer": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CursorParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by ticket status.",
            "schema": { "type": "string", "enum": ["open", "in_progress", "resolved", "closed"] }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Full-text search term (max 200 characters).",
            "schema": { "type": "string", "maxLength": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of tickets belonging to the caller's organisation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "nextCursor"],
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/PortalTicketListItem" } },
                    "nextCursor": { "type": ["string", "null"] }
                  }
                },
                "example": {
                  "data": [{
                    "id": "aa000000-0000-0003-0000-000000000001",
                    "reference": "TKT-0001",
                    "subject": "Login issue on mobile app",
                    "status": "open",
                    "priority": "P2",
                    "categoryPath": null,
                    "createdAt": "2026-01-15T10:00:00Z",
                    "updatedAt": "2026-01-15T10:00:00Z",
                    "sla": {
                      "firstResponseTargetAt": "2026-01-15T14:00:00Z",
                      "resolutionTargetAt": "2026-01-17T10:00:00Z",
                      "state": "running"
                    }
                  }],
                  "nextCursor": null
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "post": {
        "operationId": "createPortalTicket",
        "summary": "Submit a new support request from the portal",
        "tags": ["portal-tickets"],
        "security": [{ "PortalBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePortalTicketRequest" },
              "example": {
                "subject": "Login issue on mobile app",
                "description": "Cannot login after updating the app to v4.2.1.",
                "requestedPriority": "P2"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "traceId"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/PortalTicketDetail" },
                    "traceId": { "$ref": "#/components/schemas/TraceId" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/portal/tickets/{id}": {
      "get": {
        "operationId": "getPortalTicket",
        "summary": "Get portal ticket detail with public comments and SLA projection",
        "tags": ["portal-tickets"],
        "security": [{ "PortalBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "Ticket UUID.", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Ticket detail with public comments, SLA projection and status history.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PortalTicketDetail" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/portal/tickets/{id}/comments": {
      "post": {
        "operationId": "addPortalComment",
        "summary": "Add a public comment to a portal ticket",
        "description": "Visibility is always forced to `public`. Supplying a `visibility` field returns 400. Replies to closed tickets return 422 TICKET_CLOSED unless the tenant policy permits re-open.",
        "tags": ["portal-tickets"],
        "security": [{ "PortalBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "description": "Ticket UUID.", "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddPortalCommentRequest" },
              "example": { "body": "I tried restarting — still broken." }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added successfully.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PortalComment" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/portal/attachments/{id}/download": {
      "get": {
        "operationId": "downloadPortalAttachment",
        "summary": "Get a pre-signed download URL for a portal attachment",
        "tags": ["portal-attachments"],
        "security": [{ "PortalBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Pre-signed download URL valid for 5 minutes.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AttachmentDownload" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/api/v1/tickets": {
      "post": {
        "operationId": "createTicket",
        "summary": "Create a new ticket (agent)",
        "tags": ["agent-tickets"],
        "security": [{ "StaffBearer": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTicketRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Ticket created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "traceId"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Ticket" },
                    "traceId": { "$ref": "#/components/schemas/TraceId" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "get": {
        "operationId": "listTickets",
        "summary": "List / queue tickets with saved-view filtering and cursor pagination",
        "tags": ["agent-tickets"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CursorParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "name": "viewId", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "required": false, "schema": { "type": "string", "enum": ["open", "in_progress", "resolved", "closed"] } }
        ],
        "responses": {
          "200": {
            "description": "Paginated ticket list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "nextCursor"],
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Ticket" } },
                    "nextCursor": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/tickets/{id}": {
      "get": {
        "operationId": "getTicket",
        "summary": "Get a ticket by ID",
        "tags": ["agent-tickets"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Ticket detail.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": { "data": { "$ref": "#/components/schemas/Ticket" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "patch": {
        "operationId": "updateTicket",
        "summary": "Update ticket fields (partial update)",
        "description": "Optimistic concurrency: include the current `version` from a prior GET in the `If-Match` header or `version` body field. Returns 409 on stale version.",
        "tags": ["agent-tickets"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTicketRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Ticket updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": { "data": { "$ref": "#/components/schemas/Ticket" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/api/v1/tickets/{id}/comments": {
      "post": {
        "operationId": "addComment",
        "summary": "Add a comment to a ticket",
        "tags": ["agent-comments"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddCommentRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Comment added.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "get": {
        "operationId": "listComments",
        "summary": "List comments on a ticket",
        "tags": ["agent-comments"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "$ref": "#/components/parameters/CursorParam" },
          { "$ref": "#/components/parameters/LimitParam" }
        ],
        "responses": {
          "200": {
            "description": "Paginated comment list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "nextCursor"],
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } },
                    "nextCursor": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/api/v1/attachments/{id}/download": {
      "get": {
        "operationId": "downloadAttachment",
        "summary": "Get a pre-signed download URL for an agent attachment",
        "tags": ["agent-attachments"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Pre-signed download URL valid for 5 minutes.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AttachmentDownload" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TraceId": {
        "type": "string",
        "description": "Correlation ID echoed from X-Trace-ID request header.",
        "example": "01HV2MXPQ3Y5HRGZTBJE8DKNN4"
      },
...
      "TicketStatus": { "type": "string", "enum": ["open", "in_progress", "resolved", "closed"], "description": "Ticket lifecycle status." },
      "TicketPriority": { "type": "string", "enum": ["P1", "P2", "P3", "P4"], "description": "SLA priority (P1 = critical, P4 = low)." },
...
      "CreateTicketRequest": {
        "type": "object",
        "required": ["subject", "organization_id"],
        "additionalProperties": false,
        "properties": {
          "subject": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string", "maxLength": 100000 },
          "priority": { "$ref": "#/components/schemas/TicketPriority" },
          "organization_id": { "type": "string", "format": "uuid" },
          "requester_contact_id": { "type": "string", "format": "uuid" },
          "category_id": { "type": "string", "format": "uuid" }
        }
      },
      "UpdateTicketRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "subject": { "type": "string", "minLength": 1, "maxLength": 255 },
          "priority": { "$ref": "#/components/schemas/TicketPriority" },
          "status": { "$ref": "#/components/schemas/TicketStatus" },
          "assigneeId": { "type": "string", "format": "uuid" },
          "version": { "type": "integer" }
        }
      },
...
      "PortalTicketListItem": {
        "type": "object",
        "required": ["id", "subject", "status", "priority", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "reference": { "type": ["string", "null"] },
          "subject": { "type": "string" },
          "status": { "$ref": "#/components/schemas/TicketStatus" },
          "priority": { "$ref": "#/components/schemas/TicketPriority" },
          "categoryPath": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "sla": { "oneOf": [{ "$ref": "#/components/schemas/PortalSlaProjection" }, { "type": "null" }] }
        }
      },
...
      "PortalComment": {
        "type": "object",
        "required": ["id", "body", "authorDisplayName", "authorType", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "body": { "type": "string" },
          "authorDisplayName": { "type": "string" },
          "authorType": { "type": "string", "enum": ["customer", "agent"] },
          "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/PortalAttachmentMeta" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PortalTicketDetail": {
        "type": "object",
        "required": ["id", "subject", "status", "priority", "createdAt", "updatedAt", "comments", "statusHistory"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "reference": { "type": ["string", "null"] },
          "subject": { "type": "string" },
          "status": { "$ref": "#/components/schemas/TicketStatus" },
          "priority": { "$ref": "#/components/schemas/TicketPriority" },
          "categoryPath": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "sla": { "oneOf": [{ "$ref": "#/components/schemas/PortalSlaProjection" }, { "type": "null" }] },
          "aiSummary": { "type": "string" },
          "comments": { "type": "array", "items": { "$ref": "#/components/schemas/PortalComment" } },
          "statusHistory": { "type": "array", "items": { "$ref": "#/components/schemas/PortalStatusHistoryEntry" } }
        }
      },
      "CreatePortalTicketRequest": {
        "type": "object",
        "required": ["subject", "description"],
        "additionalProperties": false,
        "properties": {
          "subject": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "minLength": 1, "maxLength": 20000 },
          "categoryId": { "type": "string", "format": "uuid" },
          "requestedPriority": { "$ref": "#/components/schemas/TicketPriority" },
          "attachmentIds": { "type": "array", "items": { "type": "string", "format": "uuid" }, "maxItems": 10 }
        }
      },
      "AddPortalCommentRequest": {
        "type": "object",
        "required": ["body"],
        "additionalProperties": false,
        "description": "visibility field MUST NOT be present.",
        "properties": {
          "body": { "type": "string", "minLength": 1, "maxLength": 20000 },
          "attachmentIds": { "type": "array", "items": { "type": "string", "format": "uuid" }, "maxItems": 10 }
        }
      },
...
      "AttachmentDownload": {
        "type": "object",
        "required": ["url", "expiresAt"],
        "properties": {
          "url": { "type": "string", "format": "uri" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "CursorParam": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque base64url-encoded continuation token. Omit to start from the first page.",
        "schema": { "type": "string" }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Number of items to return per page. Maximum 100.",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request — validation failed or malformed input.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid bearer token.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Forbidden": {
        "description": "Forbidden — insufficient permissions.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "NotFound": {
        "description": "Not Found — resource does not exist or is out of scope.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Conflict": {
        "description": "Conflict — optimistic concurrency version mismatch.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "UnprocessableEntity": {
        "description": "Unprocessable Entity — business rule violation.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "TooManyRequests": {
        "description": "Too Many Requests — rate limit exceeded.",
        "headers": {
          "Retry-After": {
            "description": "Number of seconds to wait before retrying.",
            "schema": { "type": "integer", "example": 60 }
          }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      }
    },
    "securitySchemes": {
      "StaffBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC JWT issued by the tenant identity provider (Okta / Entra ID). Staff and agent principals."
      },
      "PortalBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Portal-scoped JWT bound to a single customer organisation."
      },
      "MachineToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Long-lived tenant-issued API key for machine-to-machine integrations."
      }
    }
  }
}